+++ data/pace.csv
@@ -1,3 +1,4 @@
+userid,total_time,distance
 1,25,5000
 2,24,4800
 3,28,6000
