+++ data/users.csv
@@ -1,0 +1,15 @@
+userid,username,age
+1,user1,24
+2,ahmet,21
+3,yunus,28
+5,john,34
+6,adam,31
+7,brian,41
+8,tyler,26
+9,ceren,35
+10,hasan,43
+11,selen,29
+12,diana,45
+13,thomas,44
+14,james,39
+15,janet,36
