--- nginx/nginx.conf
@@ -1,16 +1,0 @@
-upstream web_api {
-    server web:5000;
-}
-
-server {
-
-    listen 80;
-
-    location / {
-        proxy_pass http://web_api;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-        proxy_set_header Host $host;
-        proxy_redirect off;
-    }
-
-}
