--- web/.env.prod
@@ -1,3 +1,0 @@
-FLASK_APP=project/__init__.py
-FLASK_ENV=production
-DATABASE_URL=postgresql://admin:password@database:5432/database
