--- web/Dockerfile
@@ -1,24 +1,0 @@
-# pull official base image
-FROM python:3.8.0-alpine
-
-# set work directory
-WORKDIR /usr/src/app
-
-# set environment variables
-ENV PYTHONDONTWRITEBYTECODE 1
-ENV PYTHONUNBUFFERED 1
-
-RUN apk update && apk add postgresql-dev gcc python3-dev musl-dev
-
-# install dependencies
-RUN pip install --upgrade pip
-COPY ./requirements.txt /usr/src/app/requirements.txt
-RUN pip install -r requirements.txt
-
-# copy project
-COPY . /usr/src/app/
-
-EXPOSE 5000
-
-ENTRYPOINT ["/usr/src/app/entrypoint.sh"]
-CMD ["python", "manage.py", "run", "-h", "0.0.0.0"]
