--- web/data/users.csv
@@ -1,15 +1,0 @@
-1,user1,24,gender queer
-2,ahmet,21,male
-3,yunus,28,male
-4,ilke,25,male
-5,john,34,male
-6,adam,31,male
-7,brian,41,male
-8,tyler,26,male
-9,ceren,35,female
-10,hasan,43,male
-11,selen,29,female
-12,diana,45,female
-13,thomas,44,male
-14,james,39,male
-15,janet,36,female
