--- web/entrypoint.sh
@@ -1,1 +1,0 @@
-#!/bin/sh
