--- web/manage.py
@@ -1,3 +1,0 @@
-from project import create_app
-
-app = create_app()
