+++ web/requirements.txt
@@ -1,29 +1,6 @@
-alembic==1.4.3
-attrs==20.2.0
 click==7.1.2
-coverage==5.3
 Flask==1.1.2
-flask-marshmallow==0.14.0
-Flask-Migrate==2.5.2
-Flask-SQLAlchemy==2.4.4
-gunicorn==20.0.4
-iniconfig==1.0.1
 itsdangerous==1.1.0
 Jinja2==2.11.2
-Mako==1.1.3
 MarkupSafe==1.1.1
-marshmallow==3.8.0
-marshmallow-sqlalchemy==0.23.1
-packaging==20.4
-pluggy==0.13.1
-psycopg2-binary==2.8.6
-py==1.9.0
-pyparsing==2.4.7
-pytest==6.1.1
-python-dateutil==2.8.1
-python-dotenv==0.14.0
-python-editor==1.0.4
-six==1.15.0
-SQLAlchemy==1.3.19
-toml==0.10.1
 Werkzeug==1.0.1
