--- web/tests/conftest.py
@@ -1,11 +1,0 @@
-import pytest
-from project import create_app
-
-@pytest.fixture
-def app():
-    app = create_app(testing=True)
-    yield app
-
-@pytest.fixture
-def client(app):
-    return app.test_client()
